:art: Remove create-tenant "extra settings" example (#1732)
This change removes the existing example that was used in the `extra_settings` field for the CreateTenant endpoint.

The example was there in order to illustrate an advanced-user setting (changing `ACAPY_AUTO_ACCEPT_INVITES` to False), but this has the drawback of that example being auto-filled, in the Scalar UI, in the sample create tenant request payload. Meaning, if someone naively creates a tenant with the example request, then the auto-accept flags would be disabled, breaking the recommended flows. Rather than changing the example to set the auto-accept value to True, it's better to just remove it, and avoid all confusion, since True is already the default behaviour.

The extra_settings field description is also updated to indicate that it is only relevant for advanced use cases.